Ensure "messages" or "updates" input is always a list of objects

Utility to normalize webhook payloads so downstream parsers can safely map_dfr(). Accepts single objects, arrays of objects, or NULL, and always returns a list.

Usage

whapi_as_array(x, kind = c("message", "update"))

Arguments

x

Input object: NULL, a list (single message/update), or a list of lists.

kind

Either "message" or "update". Defines which keys to look for when deciding if input is a single object.

Value

A list of message/update objects (possibly length 0).

Examples

whapi_as_array(list(id="abc", type="text"), kind="message")
#> [[1]]
#> [[1]]$id
#> [1] "abc"
#> 
#> [[1]]$type
#> [1] "text"
#> 
#> 
whapi_as_array(NULL, "update") # empty list
#> list()